requirements.txt
Django==2.0.4
djangorestframework==3.8.0
pytest
pytest-django
pytest-cov
django-rest-swagger==2.1.2
mixer
psycopg2-binary==2.8.3
sentry-sdk==0.10.2
django-filter
pandas
